wget.sh: skip checksum verification if ${SKIP_CHECKSUM} set to 1

+++ scripts/wget.sh
@@ -83,23 +83,25 @@ if [ ! -f $file ]; then
 	exit 1
 fi
 
-if [ ! -f "${TOP}/checksums/$(basename ${file}).sha512" ]; then
-	echo "Missing checksum for $(basename ${file})" >&2
-	if [ "x${CREATE_CHECKSUM}" == "x1" ]; then
-		echo "Creating ${TOP}/checksums/$(basename ${file}).sha512 as requested" >&2
-		(sha512sum $file | cut -d ' ' -f1 > ${TOP}/checksums/$(basename ${file}).sha512) || exit 1
-		exit 0
+if [ "x${SKIP_CHECKSUM}" != "x1" ]; then
+	if [ ! -f "${TOP}/checksums/$(basename ${file}).sha512" ]; then
+		echo "Missing checksum for $(basename ${file})" >&2
+		if [ "x${CREATE_CHECKSUM}" == "x1" ]; then
+			echo "Creating ${TOP}/checksums/$(basename ${file}).sha512 as requested" >&2
+			(sha512sum $file | cut -d ' ' -f1 > ${TOP}/checksums/$(basename ${file}).sha512) || exit 1
+			exit 0
+		fi
+		echo "Removing ${file}" >&2
+		rm -f ${file}
+		exit 1
 	fi
-	echo "Removing ${file}" >&2
-	rm -f ${file}
-	exit 1
-fi
 
-if [ "`sha512sum $file | cut -d ' ' -f1`" != "`cat ${TOP}/checksums/$(basename ${file}).sha512`" ]; then
-	echo "Checksum mismatch" >&2
-	echo "Removing ${file}" >&2
-	rm -f ${file}
-	exit 1
+	if [ "`sha512sum $file | cut -d ' ' -f1`" != "`cat ${TOP}/checksums/$(basename ${file}).sha512`" ]; then
+		echo "Checksum mismatch" >&2
+		echo "Removing ${file}" >&2
+		rm -f ${file}
+		exit 1
+	fi
 fi
 
 exit 0
